+++ ship-mate.el
@@ -84,48 +84,27 @@ nil, show immediately."
   :group 'ship-mate
   :type 'integer)
 
-(defcustom ship-mate-subcommands-key "x"
-  "The key used to bind subcommands.
-
-If this is nil, they aren't bound."
-  :group 'ship-mate
-  :type '(choice (key :tag "The key to bind to")
-                 (const :tag "Don't bind" nil)))
-
 ;;;; Variables
-
-(defvar ship-mate-subcommand-map
-  (let ((map (make-sparse-keymap)))
-
-    (define-key map "r" #'ship-mate-hidden-recompile)
-    (define-key map "h" #'ship-mate-hide-visible)
-    (define-key map "c" #'ship-mate-select-command)
-
-    (define-key map "s" #'ship-mate-show-results)
-    (define-key map "S" #'ship-mate-show-hidden)
-
-    (define-key map "e" #'ship-mate-edit-history)
-    (define-key map "E" #'ship-mate-refresh-history)
-    (define-key map "n" #'ship-mate-edit-environment)
-
-    map)
-  "Command map for additional `ship-mate' commands.")
 
 (defvar ship-mate-command-map
   (let ((map (make-sparse-keymap)))
 
-    (when ship-mate-subcommands-key
-      (define-key
-       map
-       ship-mate-subcommands-key
-       `("subcommands" . ,ship-mate-subcommand-map)))
+    (define-key map (kbd ".") #'ship-mate-select-command)
+    (define-key map (kbd ",") #'ship-mate-hidden-recompile)
+    (define-key map (kbd "/") #'ship-mate-hide-visible)
+
+    (define-key map (kbd "!") #'ship-mate-show-results)
+    (define-key map (kbd "@") #'ship-mate-show-hidden)
+
+    (define-key map (kbd "%") #'ship-mate-edit-history)
+    (define-key map (kbd "^") #'ship-mate-refresh-history)
+    (define-key map (kbd "&") #'ship-mate-edit-environment)
 
     map)
   "Command map for `ship-mate' commands.
 
-Each command created by `ship-mate-create-command' will be
-automatically bound in here using either the provided key or the
-initial of the command's name.")
+Commands created by `ship-mate-create-command' are automatically
+bound here using an uppercase letter.")
 
 (defvar ship-mate-commands nil
   "List of commands and their per-project histories.